requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0
requests>=2.31
pypdf>=4.0          # PDF text extraction
python-docx>=1.1    # DOCX text extraction
odfpy>=1.4          # ODT text extraction
